pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "axa-fr-app-settings"
version = "0.1.0"
description = "Typed application settings with a .NET-like configuration builder for Python."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
  { name = "AXA France" }
]
keywords = ["configuration", "settings", "pydantic", "yaml", "dotnet"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
dependencies = [
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0,<7.0",
  "python-dotenv>=1.0,<2.0",
]

[project.urls]
Homepage = "https://github.com/your-org/axa-fr-app-settings"
Repository = "https://github.com/your-org/axa-fr-app-settings"
Issues = "https://github.com/your-org/axa-fr-app-settings/issues"

[dependency-groups]
dev = [
  "pytest>=8.2,<9.0",
  "ruff>=0.9.0,<1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/axa_fr_app_settings"]

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
  "/examples",
  "/README.md",
  "/LICENSE",
  "/pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
